pyproject.toml
[project]
name = "clawgate"
version = "0.1.0"
description = "Agent-Aware LLM Gateway with Smart Routing, Context Management, and Priority Scheduling"
authors = [
    {name = "OpenClaw Team", email = "noreply@openclaw.ai"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # 推理引擎
    "mlx-lm>=0.10.0; platform_machine=='arm64' and sys_platform=='darwin'",
    "llama-cpp-python>=0.2.50",

    # 数据库
    "aiosqlite>=0.19.0",
    "tantivy>=0.21.0",

    # HTTP 客户端
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",

    # 工具库
    "tiktoken>=0.5.2",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # 监控
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
